+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/engine/middleware/EngineDelegateMiddleware.kt
@@ -41,6 +41,7 @@ internal class EngineDelegateMiddleware(
             is EngineAction.GoToHistoryIndexAction -> goToHistoryIndex(context.store, action)
             is EngineAction.ToggleDesktopModeAction -> toggleDesktopMode(context.store, action)
             is EngineAction.ExitFullScreenModeAction -> exitFullScreen(context.store, action)
+            is EngineAction.SaveToPdfAction -> saveToPdf(context.store, action)
             is EngineAction.ClearDataAction -> clearData(context.store, action)
             is EngineAction.PurgeHistoryAction -> purgeHistory(context.state)
             else -> next(action)
@@ -131,6 +132,14 @@ internal class EngineDelegateMiddleware(
     ) = scope.launch {
         getEngineSessionOrDispatch(store, action)
             ?.exitFullScreenMode()
+    }
+
+    private fun saveToPdf(
+        store: Store<BrowserState, BrowserAction>,
+        action: EngineAction.SaveToPdfAction
+    ) = scope.launch {
+        getEngineSessionOrDispatch(store, action)
+            ?.requestPdfToDownload()
     }
 
     private fun clearData(