+++ mobile/android/android-components/docs/changelog.md
@@ -17,6 +17,12 @@ permalink: /changelog/
 * **feature-prompts**:
   * Added prompt dismiss listener to `ChoicePromptDelegate`. [#12562](https://github.com/mozilla-mobile/android-components/issues/12562)
 
+* **browser-engine-gecko**:
+  * Add support for Save to PDF in the
+
+* **concept-engine**:
+  * Add new `requestPdfToDownload` API in `Engine`. This is currently only supported in the Gecko Engine.
+
 * **browser-storage-sync**:
   * Stop reporting to the crash servers the expected `OperationInterrupted` exceptions for when interrupting in progress reads/writes from Application-Services. [#12557](https://github.com/mozilla-mobile/android-components/issues/12557), [#12569](https://github.com/mozilla-mobile/android-components/issues/12569).
 
@@ -42,7 +48,7 @@ permalink: /changelog/
 
 * **browser-toolbar**
   * 🚒 Bug fixed [issue #12497](https://github.com/mozilla-mobile/android-components/issues/12497) - Set the same margin to toolbar background for display as for edit
-  * 
+
 * **feature-recentlyclosed**
   * 🚒 Bug fixed [issue #12470](https://github.com/mozilla-mobile/android-components/issues/12470) - Set autoMirrored to true to fix RTL issues
 
@@ -69,7 +75,7 @@ permalink: /changelog/
 * [Configuration](https://github.com/mozilla-mobile/android-components/blob/v103.0.0/.config.yml)
 
 * **site-permission-feature**
-  * 🆕 [issue #12345](https://github.com/mozilla-mobile/android-components/issues/12345) 
+  * 🆕 [issue #12345](https://github.com/mozilla-mobile/android-components/issues/12345)
   * - Add an id for all site permissions allowing to easily identify with what permissions the user interacts.
   * - Emit facts for when a permission prompt is shown/allowed/denied.